refact!: Panel request controllers
