+++ src/Panel/Controller/Controller.php
@@ -3,6 +3,7 @@
 namespace Kirby\Panel\Controller;
 
 use Kirby\Cms\App;
+use Kirby\Cms\Site;
 use Kirby\Http\Request;
 
 /**
@@ -20,11 +21,13 @@ abstract class Controller
 {
 	protected App $kirby;
 	protected Request $request;
+	protected Site $site;
 
 	public function __construct()
 	{
 		$this->kirby   = App::instance();
 		$this->request = $this->kirby->request();
+		$this->site    = $this->kirby->site();
 	}
 
 	abstract public function load(): mixed;
